refactor(core): split router.rs into router/ submodule by responsibility
Cooldown management → router/cooldown.rs, error classification/failover →
router/failover.rs, LlmRouter dispatch → router/mod.rs. Logic unchanged.
